+++ src/linter/entity/checkWeakEntityRelationshipExists.ts
@@ -1,5 +1,5 @@
-import { ER } from "../types/parser/ER";
-import { WeakEntityDependsOnNonExistentRelationshipError } from "../types/linter/SemanticError";
+import { ER } from "../../types/parser/ER";
+import { WeakEntityDependsOnNonExistentRelationshipError } from "../../types/linter/SemanticError";
 
 export const checkWeakEntityRelationshipExists = (
   er: ER,
