+++ src/linter/index.ts
@@ -1,14 +1,14 @@
 import { SemanticError } from "../types/linter/SemanticError";
 import { ER } from "../types/parser/ER";
-import { checkEntityDuplicate } from "./checkEntityDuplicate";
-import { checkEntityDuplicateAttribute } from "./checkEntityDuplicateAttribute";
-import { checkEntityNoPrimaryKey } from "./checkEntityNoPrimaryKey";
-import { checkEntityExtendsNonExistentEntity } from "./checkExtendsExistingEntity";
-import { checkWeakEntityRelationshipExists } from "./checkWeakEntityRelationshipExists";
-import { checkWeakEntityInRelationship } from "./checkWeakEntityInRelationship";
-import { checkWeakEntityHasTotalParticipation } from "./checkWeakEntityHasTotalParticipation";
-import { checkRelationshipDuplicate } from "./checkRelationshipDuplicate";
-import { checkChildEntityHasKey } from "./checkChildEntityHasKey";
+import { checkEntityDuplicate } from "./entity/checkEntityDuplicate";
+import { checkEntityDuplicateAttribute } from "./entity/checkEntityDuplicateAttribute";
+import { checkEntityNoPrimaryKey } from "./entity/checkEntityNoPrimaryKey";
+import { checkEntityExtendsNonExistentEntity } from "./entity/checkExtendsExistingEntity";
+import { checkWeakEntityRelationshipExists } from "./entity/checkWeakEntityRelationshipExists";
+import { checkWeakEntityInRelationship } from "./entity/checkWeakEntityInRelationship";
+import { checkWeakEntityHasTotalParticipation } from "./entity/checkWeakEntityHasTotalParticipation";
+import { checkChildEntityHasKey } from "./entity/checkChildEntityHasKey";
+import { checkRelationshipDuplicate } from "./relationship/checkRelationshipDuplicate";
 
 type checkErrorFunction = (er: ER) => SemanticError[];
